alx_travel_app/requirement.txt
django
djangorestframework
django-cors-headers
drf-yasg
celery
mysqlclient
django-environ